+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 > The format is based on [Keep a Changelog](https://keepachangelog.com/de/1.0.0/).
 
 ---
+## [2.0.2] - 2026-03-14
+### 🐛 Fixed
+- Fix DHCP scope check issue with datatype
+
 ## [2.0.1] - 2026-01-12
 ### 🔄 Changed
 - License Check